Enforce Accommodation Management user permissions

+++ app/Http/Controllers/AccommodationController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Accommodation;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
+use Illuminate\Support\Facades\Gate;
 
 class AccommodationController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-accommodations');
         $accommodations = Accommodation::all();
         $title = "Accommodations";
         return view('accommodations', compact('accommodations', 'title'));
@@ -17,6 +19,7 @@ class AccommodationController extends Controller
     
     public function store(Request $request)
     {
+        Gate::authorize('create-accommodations');
         Accommodation::create($request->validate([
             'name' => 'required|unique:accommodations,name|max:255',
             'description' => 'nullable|string',
@@ -27,6 +30,7 @@ class AccommodationController extends Controller
 
     public function update(Request $request, Accommodation $accommodation)
     {
+        Gate::authorize('edit-accommodations');
         $accommodation->update($request->validate([
             'name' => 'required|max:255|unique:accommodations,name,' . $accommodation->id,
             'description' => 'nullable|string',
@@ -37,6 +41,7 @@ class AccommodationController extends Controller
 
     public function destroy(Accommodation $accommodation)
     {
+        Gate::authorize('delete-accommodations');
         try {
             $accommodation->delete();
             return back()->with('success', 'Accommodation Deleted Successfully!');
